requirements.txt
Flask==2.1.1
celery==5.2.3
redis==4.1.2
pymysql==1.0.2
psycopg2-binary==2.9.3
boto3==1.20.21
gunicorn==20.1.0
